+++ src/main/java/tw/kane/osu4j/User.java
@@ -5,7 +5,6 @@ import org.json.JSONObject;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.List;
 
 public class User {
     private final JSONObject object;
@@ -39,42 +38,42 @@ public class User {
     }
 
     private class Count {
-        int _50, _100, _300, SSH, SS, SH, S, A, plays, secondPlayed;
+        long _50, _100, _300, SSH, SS, SH, S, A, plays, secondPlayed;
 
         public Count() {
-            _50 = object.isNull("count50") ? 0 : Integer.parseInt(object.getString("count50"));
-            _100 = object.isNull("count100") ? 0 : Integer.parseInt(object.getString("count100"));
-            _300 = object.isNull("count300") ? 0 : Integer.parseInt(object.getString("count300"));
+            _50 = object.isNull("count50") ? 0 : Long.parseLong(object.getString("count50"));
+            _100 = object.isNull("count100") ? 0 : Long.parseLong(object.getString("count100"));
+            _300 = object.isNull("count300") ? 0 : Long.parseLong(object.getString("count300"));
 
-            SSH = object.isNull("count_rank_ssh") ? 0 : Integer.parseInt(object.getString("count_rank_ssh"));
-            SS = object.isNull("count_rank_ss") ? 0 : Integer.parseInt(object.getString("count_rank_ss"));
-            SH = object.isNull("count_rank_sh") ? 0 : Integer.parseInt(object.getString("count_rank_sh"));
-            S = object.isNull("count_rank_s") ? 0 : Integer.parseInt(object.getString("count_rank_s"));
-            A = object.isNull("count_rank_a") ? 0 : Integer.parseInt(object.getString("count_rank_a"));
+            SSH = object.isNull("count_rank_ssh") ? 0 : Long.parseLong(object.getString("count_rank_ssh"));
+            SS = object.isNull("count_rank_ss") ? 0 : Long.parseLong(object.getString("count_rank_ss"));
+            SH = object.isNull("count_rank_sh") ? 0 : Long.parseLong(object.getString("count_rank_sh"));
+            S = object.isNull("count_rank_s") ? 0 : Long.parseLong(object.getString("count_rank_s"));
+            A = object.isNull("count_rank_a") ? 0 : Long.parseLong(object.getString("count_rank_a"));
 
-            plays = object.isNull("playcount") ? 0 : Integer.parseInt(object.getString("playcount"));
-            secondPlayed = object.isNull("total_second_played") ? 0 : Integer.parseInt(object.getString("total_seconds_played"));
+            plays = object.isNull("playcount") ? 0 : Long.parseLong(object.getString("playcount"));
+            secondPlayed = object.isNull("total_second_played") ? 0 : Long.parseLong(object.getString("total_seconds_played"));
         }
     }
 
     private class PlayScore {
-        int ranked, total;
+        long ranked, total;
 
         public PlayScore() {
-            ranked = object.isNull("ranked_score") ? 0 : Integer.parseInt(object.getString("ranked_score"));
-            total = object.isNull("total_score") ? 0 : Integer.parseInt(object.getString("total_score"));
+            ranked = object.isNull("ranked_score") ? 0 : Long.parseLong(object.getString("ranked_score"));
+            total = object.isNull("total_score") ? 0 : Long.parseLong(object.getString("total_score"));
         }
     }
 
     private class PerformancePoint {
         float pp;
-        int rank, countryRank;
+        long rank, countryRank;
 
         public PerformancePoint() {
             pp = object.isNull("pp_raw") ? 0 : Float.parseFloat(object.getString("pp_raw"));
 
-            rank = object.isNull("pp_rank") ? 0 : Integer.parseInt(object.getString("pp_rank"));
-            countryRank = object.isNull("pp_country_rank") ? 0 : Integer.parseInt(object.getString("pp_country_rank"));
+            rank = object.isNull("pp_rank") ? 0 : Long.parseLong(object.getString("pp_rank"));
+            countryRank = object.isNull("pp_country_rank") ? 0 : Long.parseLong(object.getString("pp_country_rank"));
         }
     }
 }
